requirements.txt
asgiref==3.7.2
boto3==1.28.29
botocore==1.31.29
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.4
django-allauth==0.54.0
django-allauth-2fa==0.11.1
django-encrypted-fields-2==1.1.13
django-encrypted-model-fields==0.6.5
django-otp==1.2.2
django-storages==1.13.2
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
packaging==23.1
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.8.0
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-keyczar-2==0.725
python3-openid==3.2.0
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.18
whitenoise==6.5.0
bcrypt
pycryptodome
psycopg2-binary
